+++ Dozentenplanung/Controllers/SkillController.cs
@@ -19,42 +19,51 @@ namespace Dozentenplanung.Controllers
 
         public IActionResult Index()
         {
+            this.PutRolesInViewBag();
             return View("Skills", this.DatabaseContext.Skills);
         }
 
         public IActionResult Edit(int? id)
         {
-            Skill skill;
-            if (id.HasValue)
-            {
-                skill = this.DatabaseContext.SkillForId(id.Value);
-            }
-            else
-            {
-                SkillBuilder skillBuilder = new SkillBuilder(this.DatabaseContext);
-                skillBuilder.Title = "Skill";
-                skillBuilder.Description = "Bescheibung";
-                skillBuilder.Save();
-                skill = skillBuilder.Skill();
+            if (this.CurrentUserIsAdministrator()) {
+                Skill skill;
+                if (id.HasValue)
+                {
+                    skill = this.DatabaseContext.SkillForId(id.Value);
+                }
+                else
+                {
+                    SkillBuilder skillBuilder = new SkillBuilder(this.DatabaseContext);
+                    skillBuilder.Title = "Skill";
+                    skillBuilder.Description = "Bescheibung";
+                    skillBuilder.Save();
+                    skill = skillBuilder.Skill();
+                }
+                return View(skill);    
+            } else {
+                return RedirectToAction("index");
             }
 
-            return View(skill);
         }
 
         public IActionResult Save(int id, string title, string description)
         {
-            SkillBuilder skillBuilder = new SkillBuilder(this.DatabaseContext, this.DatabaseContext.SkillForId(id));
-            skillBuilder.Title = title;
-            skillBuilder.Description = description;
-            skillBuilder.Save();
-            return RedirectToAction("index", "skill");
+            if (this.CurrentUserIsAdministrator()) {
+                SkillBuilder skillBuilder = new SkillBuilder(this.DatabaseContext, this.DatabaseContext.SkillForId(id));
+                skillBuilder.Title = title;
+                skillBuilder.Description = description;
+                skillBuilder.Save();    
+            }
+            return RedirectToAction("index");
         }
 
         public IActionResult Delete(int id)
         {
-            Skill skill = this.DatabaseContext.SkillForId(id);
-            skill.DeleteFromContext(this.DatabaseContext);
-            this.SaveDatabaseContext();
+            if (this.CurrentUserIsAdministrator()) {
+                Skill skill = this.DatabaseContext.SkillForId(id);
+                skill.DeleteFromContext(this.DatabaseContext);
+                this.SaveDatabaseContext();    
+            }
             return RedirectToAction("index", "skill");
         }
     }
